Change variable name
Variable named '_file' overlaps the variables in several functions in 'spxgen.sht'
It's not problem with shells having 'local' but in ksh (typeset) or empty 'local'
this '_file' variable assignment have influence at the upper level.
To prevent this the '_file' variable renamed to '_path'.
Other option is make function execution in subshell.