Add '#%ifinclude' directive
In the process of implementation of '#%ifinclude' directive some 'global'
changes was made as well:

There was the problem with ksh 'typeset' (and empty 'local' if shell
does not support that). Cause 'typeset' has no 'local' meaning to
functions defined in POSIX-style (func_name ()), its values was rewritten
in recursive invocations. So error messages didn't contain accurate information
about processed file name and line number.

One of the way to correct this is use 'function func_name' definition in ksh.
But it make the things very complicated to determine which shell is used etc.

The workaround was used to invoke 'process_file' function in subshell and it worked.

But the next problem arised: 'global' variables (such as FSTACK) can't be changed
'upward' in the calling shell environment.

In this case another way was selected: to make 'global' variables working
as expected the 'process_file' function is executed in current shell environment
again but some additional 'global' variables added in 'read_file' function
to save and restore current file name and line number (another workaround for ksh).

Well, it is probably not elegant solution but seems to work alike
in all tested shells (dash/POSIX, bash, ksh).