Simplify file processing
To clarify the processing of file two 'global' variables CURFILE and CURLINE
is used. It allows to simplify functions for directives handling.
Other change is forcing 'read_file' to be executed in subshell eliminating the need
of saving current file (RSTACK variable thus is removed).
Some sort of 'hack' is used to 'translate' PLIST variable from subshell
to upper execution environment.